Create a real test file for the pre-signup trigger
This gets the tests passing again.

Also, make the handler strict mode.
