--- lambdas/cognito-pre-signup-trigger/__tests__/demo.js
@@ -1,10 +1,0 @@
-const index = require('../index')
-
-test('should always confirm users', async () => {
-  const event = {
-    userName: 'username',
-    request: { userAttributes: {} }
-  }
-  const result = await index.handler(event)
-  expect(result).toEqual(event)
-})
