+++ lambdas/cognito-pre-signup-trigger/index.js
@@ -1,3 +1,4 @@
+'use strict'
 // Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
